requirements.txt
confluent-kafka
fastavro
requests
six
pandas
python-dotenv
pymongo[srv]==4.1.1
jsonschema==3.2.0
dnspython
-e .